+++ src/components/POLine/POLineForm.js
@@ -32,6 +32,7 @@ import {
   useAccordionToggle,
 } from '@folio/stripes-acq-components';
 
+import { useFundDistributionValidation } from '../../common/hooks';
 import {
   isEresource,
   isPhresource,
@@ -89,6 +90,10 @@ function POLineForm({
 }) {
   const history = useHistory();
   const [hiddenFields, setHiddenFields] = useState({});
+  const { validateFundDistributionTotal } = useFundDistributionValidation(formValues);
+
+  // console.log('form', form);
+  // console.log('form state', form.getState());
 
   const identifierTypes = getIdentifierTypesForSelect(parentResources);
   const locations = parentResources?.locations?.records;
@@ -480,6 +485,7 @@ function POLineForm({
                         fundDistribution={fundDistribution}
                         name="fundDistribution"
                         totalAmount={estimatedPrice}
+                        validateFundDistributionTotal={validateFundDistributionTotal}
                       />
                     </Accordion>
                     <Accordion