+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add `New invoice` action for PO actions menu. Refs UIOR-985.
 * New invoice action (PO details). Refs UIOR-990.
+* Use BE validation of fund distributions total. Refs UIOR-978.
 
 ## [3.2.0](https://github.com/folio-org/ui-orders/tree/v3.2.0) (2022-07-07)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v3.1.2...v3.2.0)